:lipstick: Beauty code
